+++ mythic_container/MythicGoRPC/__init__.py
@@ -6,6 +6,7 @@ from .send_mythic_rpc_apitoken_create import *
 from .send_mythic_rpc_artifact_create import *
 from .send_mythic_rpc_artifact_search import *
 from .send_mythic_rpc_c2_update_status import *
+from .send_mythic_rpc_chat_channel_metadata_update import *
 from .send_mythic_rpc_callback_add_command import *
 from .send_mythic_rpc_callback_create import *
 from .send_mythic_rpc_callback_decrypt_bytes import *