+++ mythic_container/MythicGoRPC/send_mythic_rpc_task_create_subtask_group.py
@@ -8,25 +8,33 @@ class MythicRPCTaskCreateSubtaskGroupTasks:
     def __init__(self,
                  SubtaskCallbackFunction: str = None,
                  CommandName: str = None,
+                 PayloadTypeName: str = None,
                  Params: str = None,
                  ParameterGroupName: str = None,
                  Token: int = None,
+                 ResolveTaskReferences: bool = None,
                  **kwargs):
         self.SubtaskCallbackFunction = SubtaskCallbackFunction
         self.CommandName = CommandName
+        self.PayloadTypeName = PayloadTypeName
         self.Params = Params
         self.ParameterGroupName = ParameterGroupName
         self.Token = Token
+        self.ResolveTaskReferences = ResolveTaskReferences
         for k, v in kwargs.items():
             logger.info(f"Unknown kwarg {k} - {v}")
+
     def to_json(self):
         return {
             "subtask_callback_function": self.SubtaskCallbackFunction,
             "command_name": self.CommandName,
+            "payload_type_name": self.PayloadTypeName,
             "params": self.Params,
             "parameter_group_name": self.ParameterGroupName,
-            "token": self.Token
+            "token": self.Token,
+            "resolve_task_references": self.ResolveTaskReferences,
         }
+
 
 class MythicRPCTaskCreateSubtaskGroupMessage:
     def __init__(self,